feat: added rhsa curate cmd and normal output format
